all the gsp names extracted from dashboard
