+++ ukpn/load/__init__.py
@@ -1,2 +1,2 @@
 """Load the data pipes and functions"""
-from ukpn.load.power_data.grafana import automate_csv_download
+from ukpn.load.power_data.grafana_data_download import automate_csv_download
