+++ ukpn/load/power_data/__init__.py
@@ -1,1 +1,2 @@
 """Load Iter Datapipes"""
+from .grafana_data_download import automate_csv_download
